+++ src/cli/commands/interfaces/сli-command.interface.ts
