+++ src/main.ts
@@ -1,7 +1,7 @@
 import 'reflect-metadata';
 import { Container } from 'inversify';
 import Application from './rest/app.js';
-import { AppComponents } from './types/appComponents.js';
+import { AppComponents } from './types/app-components.js';
 import { createRestApplicationContainer } from './rest/container.js';
 import { createUserContainer } from './modules/user/container.js';
 import { createOfferContainer } from './modules/offer/container.js';
