+++ src/modules/common/document-exists.inerface.ts
