+++ src/modules/user/container.ts
@@ -1,10 +1,10 @@
 import { Container } from 'inversify';
 import { types } from '@typegoose/typegoose';
 import { UserServiceInterface } from './interface.js';
-import { AppComponents } from '../../types/appComponents.js';
+import { AppComponents } from '../../types/app-components.js';
 import { UserEntity, UserModel } from './entity.js';
 import UserController from './controller.js';
-import { BaseController } from '../../rest/controller/baseController.js';
+import { BaseController } from '../../rest/controller/base-controller.js';
 import UserService from './service.js';
 
 export function createUserContainer() {
