+++ src/rest/app.ts
@@ -1,13 +1,13 @@
 import { LoggerInterface } from '../core/logger/logger.interface.js';
 import { ConfigInterface } from '../core/config/config.interface.js';
 import { RestSchema } from '../core/config/rest.schema.js';
-import { AppComponents } from '../types/appComponents.js';
+import { AppComponents } from '../types/app-components.js';
 import { inject, injectable } from 'inversify';
 import { getMongoConnectionString } from '../core/helpers/db.js';
 import { DatabaseClientInterface } from '../core/db.client/db.interface.js';
 import express, { Express } from 'express';
-import { ExceptionFilterInterface } from './exceptions/exeptionFilter.interface';
-import { BaseController } from './controller/baseController.js';
+import { ExceptionFilterInterface } from './exceptions/exeption-filter.interface';
+import { BaseController } from './controller/base-controller.js';
 import { AuthenticateMiddleware } from './middleware/authenticate.js';
 
 @injectable()
