+++ src/rest/exceptions/exeption-filter.interface.ts
