+++ src/rest/exceptions/http-error.ts
