+++ src/rest/middleware/authenticate.ts
@@ -3,7 +3,7 @@ import { jwtVerify } from 'jose';
 import { StatusCodes } from 'http-status-codes';
 import { createSecretKey } from 'node:crypto';
 import { MiddlewareInterface } from './middleware.interface.js';
-import { HttpError } from '../exceptions/httpError.js';
+import { HttpError } from '../exceptions/http-error.js';
 
 export const BLACK_LIST_TOKENS: Set<string> = new Set();
 export class AuthenticateMiddleware implements MiddlewareInterface {
