+++ src/rest/middleware/validate-object-id.ts
@@ -2,7 +2,7 @@ import mongoose from 'mongoose';
 import { MiddlewareInterface } from './middleware.interface.js';
 import { NextFunction, Request, Response } from 'express';
 import { StatusCodes } from 'http-status-codes';
-import { HttpError } from '../exceptions/httpError.js';
+import { HttpError } from '../exceptions/http-error.js';
 
 const { Types } = mongoose;
