+++ src/rest/types/http-method.ts
